fixed several warnings that appear with -std=gnu99 -pedantic
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@9049 d7cf8633-e32d-0410-b094-e92efae38249

+++ cl_gecko.c
@@ -325,7 +325,7 @@ OSGK_DERIVEDTYPE(OSGK_ScriptObjectTemplate);
 typedef enum OSGK_ScriptResult
 {
   srSuccess = 0,
-  srFailed = 0x80004005L /* actually NS_ERROR_FAILURE */
+  srFailed = (int)0x80004005L /* actually NS_ERROR_FAILURE */
 } OSGK_ScriptResult;
 
 typedef struct OSGK_ScriptObjectCreateParams_s
